allow user to supply spin flipper log and average value

+++ Framework/Algorithms/inc/MantidAlgorithms/DetermineSpinStateOrder.h
@@ -32,7 +32,7 @@ private:
   void exec() override;
 
   std::string m_spinFlipperLogName;
-  int m_rfStateCondition;
+  double m_rfStateCondition;
 };
 
 } // namespace Algorithms
